Refactor allocations tasks into submodules
