+++ keystone_api/plugins/slurm.py
@@ -7,6 +7,15 @@ from subprocess import PIPE, Popen
 
 log = logging.getLogger(__name__)
 
+__all__ = [
+    'get_cluster_limit',
+    'get_cluster_usage',
+    'get_slurm_account_names',
+    'get_slurm_account_principal_investigator',
+    'get_slurm_account_users',
+    'set_cluster_limit',
+]
+
 
 def subprocess_call(args: list[str]) -> str:
     """Wrapper method for executing shell commands via ``Popen.communicate``
